few modifs to avoid loops when including packages
